feat: Middleware introduced, along with the Task Managemen(Add and VIEW)t Endpoints completed

+++ pkg/domain/jwt.go
@@ -3,7 +3,7 @@ package domain
 import "github.com/golang-jwt/jwt"
 
 type JWTClaims struct {
-	Userid uint64
+	Userid uint
 	Source string
 	Role   string
 	jwt.StandardClaims
